# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API prueba previa junior PHP",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pista": {
            "get": {
                "tags": [
                    "Pista"
                ],
                "summary": "Mostrar pistas",
                "operationId": "801b02bbaba220d35fb3bdd5c7059205",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las pistas.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pista"
                ],
                "summary": "Crear pista",
                "operationId": "233c41c45e76d4a0c303d080a7dc7111",
                "requestBody": {
                    "request": "PistaStoreRequest",
                    "description": "Contiene los datos de la pista a añadir",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "pista": {
                                    "ancho": 60,
                                    "largo": 120,
                                    "deporte": "balonmano"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista creada de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pista/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Mostrar una pista",
                "operationId": "66f2bcaf98f0a22a7cdf499a9d7706cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id de la pista a mostrar",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar una pista.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pista"
                ],
                "summary": "Modificar pista",
                "operationId": "4af8fc4590069fb88626e708ba81e5d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id de la pista a modificar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "request": "PistaUpdateRequest",
                    "description": "Contiene los datos de la pista a modificar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "pista": {
                                    "ancho": 40,
                                    "largo": 100,
                                    "deporte": "baloncesto"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista modificada de manera exitosa",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pista"
                ],
                "summary": "Borrar una pista",
                "operationId": "7806d9c75defed9fa492d7bdc02cd664",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id de la pista a eliminar",
                        "required": true
                    }
                ],
                "requestBody": {
                    "request": "PistaDeleteRequest",
                    "description": "Contiene el código de la pista a elimianr",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "pista": {
                                    "code": "BASK-11"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista eliminada correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}